Cleaned up some old files and added new ones
removed TRFLP_Final_Binning_colin.R and TRFLP_class.R b/c those got
incorporated into newer files.

New files include: TRFLP_for guinea_pigs.R and TRFLPR_functions.R these
are the latest and should be used for testing. In particular
TRFLPR_functions has an important update to the lump_bins function which
now will lump bins if the last bin and second last bin need to be
merged. It was choking on this scenario because we missed it.